requirements.txt
pyrofork==2.3.45
tgcrypto>=1.2.5
motor>=3.3.2
Flask>=2.3.3
gunicorn>=21.2.0
Jinja2>=3.1.2
werkzeug>=2.3.7
itsdangerous>=2.1.2
asyncio-throttle>=1.0.2
python-dotenv>=1.0.0
